+++ src/main/java/com/lgs/eden/api/APIException.java
@@ -6,7 +6,13 @@ package com.lgs.eden.api;
 public class APIException extends Throwable {
 
     public final APIResponseCode code;
-    public final Exception e;
+    public final Throwable e;
+
+    public APIException(APIResponseCode code) {
+        super(code.name());
+        this.code = code;
+        this.e = new UnknownError();
+    }
 
     public APIException(APIResponseCode code, Exception e) {
         super(e);
